chore(context): refactor default format to context file
