+++ src/app/DateTimePicker/DateTimePicker.tsx
@@ -72,7 +72,7 @@
  * SOFTWARE.
  */
 import { TimePicker } from '@app/DateTimePicker/TimePicker';
-import { defaultDatetimeFormat } from '@app/Shared/Services/Settings.service';
+import { defaultDatetimeFormat } from '@app/Shared/DateTimeContext';
 import { useDayjs } from '@app/utils/useDayjs';
 import { Timezone } from '@i18n/datetime';
 import { isHourIn24hAM } from '@i18n/datetimeUtils';
